requirements.txt
streamlit
pandas
numpy
scipy
plotly
openpyxl
streamlit-extras
plotnine
seaborne
matplotlib